Business layer fixes

+++ BusinessLayer/Implementations/EFUserRepository.cs
@@ -17,31 +17,40 @@ namespace BusinessLayer.Implementations
         {
             _ctx = ctx;
         }
-        public Task<User> AddUserAsync(User user)
+        public async Task<User> GetUserByIdAsync(string id)
         {
-            throw new NotImplementedException();
-        }
-        public Task<ICollection<User>> GetAllUsersAsync()
-        {
-            throw new NotImplementedException();
+            try
+            {
+                var user = await _ctx.Users.FirstOrDefaultAsync(f => f.Id == id);
+                
+                if (user == null)
+                    throw new ArgumentException($"No user with ID {id} found in the database");
+
+                return user;
+            }
+            catch (Exception e)
+            {
+                throw;
+            }
         }
 
-        public async Task<User> GetUserByIdAsync(string Id)
-        {
-            return await _ctx.Users.FirstOrDefaultAsync(f => f.Id == Id);
-        }
-
-        public Task UpdateUserAsync(User user)
-        {
-            throw new NotImplementedException();
-        }
         public async Task DeleteUserAsync(User user)
         {
-            var full = _ctx.Users.Include(c => c.Comments);
-            var users = await full.FirstOrDefaultAsync(c => c.Id == user.Id);
-            _ctx.Comments.RemoveRange(users.Comments);
-            _ctx.Users.Remove(user);
-            await _ctx.SaveChangesAsync();
+            try
+            {
+                if (user == null)
+                    throw new ArgumentNullException("Cannot delete null user");
+                
+                var usersWithComments = _ctx.Users.Include(c => c.Comments);
+                var userWithComments = await usersWithComments.FirstOrDefaultAsync(c => c.Id == user.Id);
+                _ctx.Comments.RemoveRange(userWithComments.Comments);
+                _ctx.Users.Remove(user);
+                await _ctx.SaveChangesAsync();
+            }
+            catch (Exception e)
+            {
+                throw;
+            }
         }
         public void Dispose()
         {
